Move auth_auxiliars script to utils/
