+++ assistant_agent/database/tables/bigquery/users.py
@@ -1,6 +1,6 @@
 from .bq_base import BigQueryTable
 from assistant_agent.utils.gcp.bigquery import query_data, insert_rows
-from assistant_agent.auxiliars.auth_auxiliars import get_password_hash
+from assistant_agent.utils.auth_auxiliars import get_password_hash
 from assistant_agent.config import GCPConfig
 from assistant_agent.schemas import User, UserInDB
 from datetime import datetime
