+++ assistant_agent/utils/auth_auxiliars.py
